pyproject.toml
[project]
name = "portfolio-thesis-engine"
version = "0.1.0"
description = "Semi-automated portfolio management with rigorous valuation"
requires-python = ">=3.12"
authors = [{name = "Hugo Condesa"}]
license = {text = "MIT"}
readme = "README.md"

dependencies = [
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "openai>=1.50",
    "duckdb>=1.0",
    "chromadb>=0.5",
    "sqlalchemy>=2.0",
    "streamlit>=1.38",
    "typer>=0.12",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "tenacity>=9.0",
    "yfinance>=0.2.40",
    "fastapi>=0.136.1",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.6",
    "mypy>=1.11",
    "ipython>=8.26",
    "types-pyyaml>=6.0",
]

[project.scripts]
pte = "portfolio_thesis_engine.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/portfolio_thesis_engine"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests", "api/tests"]
pythonpath = ["src", "."]
addopts = "-v --cov=portfolio_thesis_engine --cov-report=term-missing"
asyncio_mode = "auto"
markers = [
    "integration: integration tests requiring external services",
    "slow: tests that take >1s",
]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = "yfinance"
ignore_missing_imports = true
